+++ src/server/devices/dachs/dachs.ts
@@ -1,14 +1,14 @@
-import { DeviceInfo, Devices, DeviceType, iBaseDevice } from '../../devices';
-import { LogLevel } from '../../../models';
+import { DeviceInfo, Devices, DeviceType, iActuator, iBaseDevice } from '../../devices';
+import { LogLevel, RoomBase } from '../../../models';
 import { DeviceCapability } from '../DeviceCapability';
-import { LogDebugType, OwnSonosDevice, ServerLogService, Utils } from '../../services';
+import { API, LogDebugType, OwnSonosDevice, ServerLogService, Utils } from '../../services';
 import _ from 'lodash';
 import { iDachsSettings } from '../../config/iDachsSettings';
 import { DachsDeviceSettings } from '../../../models/deviceSettings/dachsSettings';
 import { DachsHttpClient, DachsInfluxClient } from './lib';
 import { iFlattenedCompleteResponse } from './interfaces';
 
-export class Dachs implements iBaseDevice {
+export class Dachs implements iBaseDevice, iActuator {
   public settings: DachsDeviceSettings = new DachsDeviceSettings();
   public readonly deviceType: DeviceType = DeviceType.Dachs;
   public readonly deviceCapabilities: DeviceCapability[] = [];
@@ -16,16 +16,28 @@ export class Dachs implements iBaseDevice {
   private readonly config: iDachsSettings;
   public fetchedData: iFlattenedCompleteResponse | undefined;
   private readonly _influxClient: DachsInfluxClient | undefined;
+  private _dachsOn: boolean = false;
+  private _tempWarmWater: number = 0;
 
   public get customName(): string {
     return this.info.customName;
   }
 
+  public get actuatorOn(): boolean {
+    return this._dachsOn;
+  }
+
+  public get tempWarmWater(): number {
+    return this._tempWarmWater;
+  }
+
   public constructor(options: iDachsSettings) {
+    this.deviceCapabilities.push(DeviceCapability.actuator);
     this._info = new DeviceInfo();
     this._info.fullName = `Dachs`;
-    this._info.customName = `Dachs`;
-    this._info.allDevicesKey = `dachs`;
+    this._info.customName = `Dachs ${options.roomName}`;
+    this._info.allDevicesKey = `dachs-${options.roomName}`;
+    this._info.room = options.roomName;
     Devices.alLDevices[this._info.allDevicesKey] = this;
     if (options.influxDb) {
       this._influxClient = new DachsInfluxClient(options.influxDb);
@@ -59,10 +71,15 @@ export class Dachs implements iBaseDevice {
     return this.info.customName;
   }
 
+  public get room(): RoomBase | undefined {
+    return API.getRoom(this.info.room);
+  }
+
   public log(level: LogLevel, message: string, debugType: LogDebugType = LogDebugType.None): void {
     ServerLogService.writeLog(level, `${this.name}: ${message}`, {
       debugType: debugType,
       deviceId: this.name,
+      room: this._info.room,
       deviceName: this.name,
     });
   }
@@ -100,6 +117,52 @@ export class Dachs implements iBaseDevice {
         this._influxClient.addMeasurementToQueue(key, value ? '1' : '0');
       }
       this._influxClient.flush();
+      this._dachsOn = this.fetchedData['Hka_Mw1.usDrehzahl'] >= 1;
+      this._tempWarmWater = this.fetchedData['Hka_Mw1.Temp.sbZS_Warmwasser'] ?? 0;
+      this.persist();
     });
+  }
+
+  public persist(): void {
+    Utils.dbo?.persistActuator(this);
+  }
+
+  public setActuator(pValue: boolean, _timeout?: number, _force?: boolean): void {
+    if (!pValue || this._dachsOn) {
+      // Dachs can only be turned on, not off
+      return;
+    }
+    this.client
+      .setKeys({
+        'Stromf_Ew.Anforderung_GLT.bAktiv': '1',
+        'Stromf_Ew.Anforderung_GLT.bAnzahlModule': '1',
+      })
+      .then(() => {
+        Utils.guardedTimeout(
+          () => {
+            this.client
+              .setKeys({
+                'Stromf_Ew.Anforderung_GLT.bAktiv': '0',
+                'Stromf_Ew.Anforderung_GLT.bAnzahlModule': '0',
+              })
+              .catch((error) => {
+                this.log(LogLevel.Error, `Error while turning off Dachs: ${error}`);
+              });
+          },
+          30000,
+          this,
+        );
+      })
+      .catch((error) => {
+        this.log(LogLevel.Error, `Error while turning on Dachs: ${error}`);
+      });
+  }
+
+  public toggleActuator(_force: boolean): boolean {
+    if (!this._dachsOn) {
+      this.setActuator(true);
+      return true;
+    }
+    return false;
   }
 }